+++ docker-compose.yml
@@ -6,7 +6,11 @@ services:
     ports:
       - '9000:9000'
     user: "trainee:mercari"
+    environment:
+      REACT_APP_API_URL: "http://localhost:9000"
   react:
     build: ./typescript/simple-mercari-web
     ports:
       - '3000:3000'
+    environment:
+      FRONT_URL: "http://localhost:3000"
