+++ go/app/const.go
@@ -1,0 +1,20 @@
+package main
+
+const (
+	ImgDir = "images"
+)
+
+type Response struct {
+	Message string `json:"message"`
+}
+
+type Items struct {
+	Items []Item `json:"items"`
+}
+
+type Item struct {
+	Id        int    `json:"id"`
+	Name      string `json:"name"`
+	Category  string `json:"category"`
+	ImageName string `json:"imageName"`
+}
